Transform geometry to grob

A grob (graphical object) is the grid-package representation of a geom and is used for plotting.

gToGrob(geom, theme = NULL, ...)

Arguments

geom
[geom]
Object of class geom.

theme
[list(7)]
Visualising options; see setTheme for details.

...
instead of providing a theme, you can also determine specific graphic parameters (see gpar) separately.

Value

Depending on the provided geometry either a pointsGrob, linesGrob, polylineGrob or a polygonGrob.

Examples

coords <- data.frame(x = c(40, 70, 70, 50),
                     y = c(40, 40, 60, 70),
                     id = 1)
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window, col = "blue")

aGrob <- gToGrob(geom = aGeom)
str(aGrob)

#> List of 7
#>  $ x         :Class 'unit'  atomic [1:4] 40 70 70 50
#>   .. ..- attr(*, "valid.unit")= int 0
#>   .. ..- attr(*, "unit")= chr "npc"
#>  $ y         :Class 'unit'  atomic [1:4] 40 40 60 70
#>   .. ..- attr(*, "valid.unit")= int 0
#>   .. ..- attr(*, "unit")= chr "npc"
#>  $ id        : int [1:4] 1 1 1 1
#>  $ id.lengths: NULL
#>  $ name      : chr "aGrob"
#>  $ gp        :List of 4
#>   ..$ col : chr "black"
#>   ..$ fill: logi NA
#>   ..$ lty : chr "solid"
#>   ..$ lwd : num 3
#>   ..- attr(*, "class")= chr "gpar"
#>  $ vp        : NULL
#>  - attr(*, "class")= chr [1:3] "polygon" "grob" "gDesc"