Generate (neutral) spatial patterns

This function is still largely work in process and very experimental!

generate(model, dimensions)

Arguments

model
[list(.)]
algorithm in which the operators to generate spatial pattern models are specified. Each operator is a list iteself and includes the operator name and its arguments as sub-elements; see Examples.

dimensions
[integerish(2)]
number of columns and rows the landscape model ought to have.

Details

Spatial pattern models (SPMs), such as neutral landscape models (NLMs), are useful to study an ecological response to a set of simulated spatial patterns. Ideally, the resulting spatial patterns are strictly controlable and in many situations they ought to be neutral (i.e. not influenced by any process or bias). Here the term SPM is used because not all functions result in neutral patterns.

Many spatial pattern models have been suggested and NLMs probably constitute the most common use-case. However, often they are merely treated as "habitat models", where only the presence or abundance of habitat is simulated. Yet, "artificial patterns or processes" introduced by human disturbance and the influence of geological, atmospheric and ecological dynamics are also of interest. We try to cover these cases with ...

... spatial pattern models:

spmGradient: Generate a (neutral) gradient pattern.

spmRandom: Generate a random pattern.

spmHeightmap: Generate a heightmap.

spmNoise: Generate patterns based on different kinds of noise.

... ecological processes models:

epmSucces: Let objects in the spatial model success.

epmDiversify: Let objects in the spatial model diversify.

epmPerforate: Create gaps (e.g. clearcuts) into objects of the spatial model.

epmFragment: Create fragmentation in the spatial model.

epmConnect: Create connecting elements in the spatial model.

Additionally you can use functions in other packages as operators, if they produce a 2D-lattice of spatial patterns; see for instance NLMR.

Examples

# NOT RUN {
## neutral landscape models
nlm <- list(list(operator = "spmRandom", seed = 12769),
            list(operator = "spmHeightmap", hurst = 0.4,
                 seed = 12769))

myLandscape <- generate(model = nlm, dimensions = c(300, 300), to_env = TRUE)
# }