Import spatial datasets

Meta function to import, i.e. load or download and unpack spatial datasets and other files containing spatial data.

loadData(files = NULL, layer = NULL, dataset = NULL,
  localPath = NULL, verbose = FALSE, ...)

Arguments

files
[character(.) | data.frame(1)]
the files, as they are called in localPath. files also accepts the (subsetted) output of catalog, i.e. a data.frame with the two columns original and abbr.

layer
[character(.) | integerish(.)]
vetor of the same length as files with name or position of the layer that shall be loaded.

dataset
[character(1)]
character vector of the dataset for which files should be imported (see obtain for all datasets); in case this is given, localPath is taken from the internal default for the respective dataset (see setPaths).

localPath
[character(1)]
the local path from where files shall be loaded. If the directory does not exist, it is created and the missing data are downloaded, given .

verbose
[logical(1)]
should additional information be printed (TRUE) or should it be suppressed (FALSE, default)?

...
[various]
other arguments of the load operators.

Value

the file to load. If there is more than one file specified, a list of those files.

Details

importData checks whether the required files are available in localPath and if this is not given, in the working directory. If nothing is found there but a dataset is given, it attempts to download the files from the respective online path (see rtPaths). Please take a look at it and in case an online resource has been altered, you can adapt it here (see setPaths).

See also

Examples

# load a bunch of '.csv'-files containing coordinates into the gloabl environment.
myLocations <- loadData(files = c("aWindow.csv", "locations.csv"),
                        localPath = system.file("csv", package="rasterTools"),
                        verbose = TRUE)

#>   |                                                                          |                                                                  |   0%

#> Parsed with column specification:
#> cols(
#>   fid = col_integer(),
#>   x = col_integer(),
#>   y = col_integer()
#> )

#>   |                                                                          |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>                                 |  50%

#> Parsed with column specification:
#> cols(
#>   id = col_integer(),
#>   fid = col_integer(),
#>   X = col_double(),
#>   Y = col_double()
#> )

#>   |                                                                          |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>| 100%