Adjacency of cells

Calculate the cell-adjacency matrix for a raster

mAdjacency(obj, type = "like", count = "double", layer = NULL)

Arguments

obj
[Raster*(1)]
The object to measure.

type
[character(1)]
which type of adjacencies to calculate; either "like" adjacencies between cells of the same value, adjacencies between "paired" values or the rowsum of the paired values ("pairedSum").

count
[character(1)]
the cells that should be counted; possible values are "single" to count only cells to the right and bottom of the focal cell and "double" to count additionally cells to the left and top of the focal cell.

layer
[character(1)]
in case obj has several layers, specify here the layer for which the adjacency shall be calculated (by default, the first layer).

Value

a data.frame of the frequency of adjacencies of the values in obj.

Details

In case type = "like", only the diagonal of the adjacency matrix is returned, in case type = "paired", the complete adjacency matrix is returned.

See also

Examples

cat <- rtData$categorical
bin <- rBinarise(rtData$continuous, thresh = 40)

# double count like adjacencies
mAdjacency(obj = cat)

#>    class likeAdj
#> 1      1     114
#> 11    11     218
#> 21    21     656
#> 24    24    3082
#> 27    27    1568
#> 31    31     242
#> 41    41    1586
#> 44    44    1418
#> 47    47    1740

# paired adjacencies
mAdjacency(obj = cat, type = "paired")

#>    class   1  11  21   24   27  31   41   44   47
#> 1      1 114   0   0    0    0   0   34    0    0
#> 11    11   0 218   0   57   36   0    2   53   47
#> 21    21   0   0 656   28    1   0   95   29    0
#> 24    24   0  57  28 3082   20   1  292   15   55
#> 27    27   0  36   1   20 1568   0   57   57   78
#> 31    31   0   0   0    1    0 242    0    0   33
#> 41    41  34   2  95  292   57   0 1586   89   77
#> 44    44   0  53  29   15   57   0   89 1418  136
#> 47    47   0  47   0   55   78  33   77  136 1740

# adjacencies with single count
mAdjacency(obj = bin, count = "single")

#>   class likeAdj
#> 0     0    3746
#> 1     1    2144