Area of objects

Calculate the area of objects in a raster

mArea(obj, scale = "patch", unit = "cells", layer = NULL)

Arguments

obj
[RasterLayer(1)]
The object to measure.

scale
[character(1)]
scale at which the area of objects should be calculated; possible values are "patch", "class" and "window".

unit
[character(1)]
the unit the output should have. With "map" the result will be in the respective map unit and with "cells" (default) it will be the number of raster cells.

layer
[character(1)]
in case obj has several layers, specify here the layer for which the area of objects shall be calculated (by default, the first layer).

Value

For scale = "window" the area of the overall raster, for scale = "class" the total area of each unique value (class), for scale = "patch" the area of distinct objects per distinct values (i.e. the area of patches per class).

See also

Examples

cat <- rtData$categorial
bin <- rBinarise(rtData$continuous, thresh = 40)

# the area ...
# ... per landuse type
mArea(obj = cat, scale = "class")

#>   class cells
#> 1     1    37
#> 2    11   104
#> 3    21   281
#> 4    24   507
#> 5    27   466
#> 6    31   407
#> 7    41   561
#> 8    44   453
#> 9    47   544

# ... of patches per landuse type
mArea(obj = cat, scale = "patch")

#>    class patch cells
#> 1      1     1    37
#> 2     11     1   104
#> 3     21     1   209
#> 4     21     2    72
#> 5     24     1   296
#> 6     24     2   173
#> 7     24     3    38
#> 8     27     1   293
#> 9     27     2   173
#> 10    31     1   259
#> 11    31     2   108
#> 12    31     3    40
#> 13    41     1    41
#> 14    41     2    59
#> 15    41     3    82
#> 16    41     4    16
#> 17    41     5    48
#> 18    41     6     2
#> 19    41     7     1
#> 20    41     8    18
#> 21    41     9     1
#> 22    41    10     1
#> 23    41    11   136
#> 24    41    12    62
#> 25    41    13    94
#> 26    44     1    38
#> 27    44     2    16
#> 28    44     3     1
#> 29    44     4    96
#> 30    44     5     1
#> 31    44     6     2
#> 32    44     7     5
#> 33    44     8   163
#> 34    44     9    99
#> 35    44    10     2
#> 36    44    11    12
#> 37    44    12     2
#> 38    44    13    10
#> 39    44    14     6
#> 40    47     1   124
#> 41    47     2     2
#> 42    47     3   136
#> 43    47     4    27
#> 44    47     5     1
#> 45    47     6     1
#> 46    47     7     1
#> 47    47     8     1
#> 48    47     9     1
#> 49    47    10     1
#> 50    47    11     1
#> 51    47    12     7
#> 52    47    13   241

# ...  of certain values
require(magrittr)
rBinarise(obj = cat, match = c(41, 44, 47)) %>%
  mArea(scale = "patch", layer = "values_binarised")

#>    class patch cells
#> 1      1     1    41
#> 2      1     2    97
#> 3      1     3   327
#> 4      1     4    16
#> 5      1     5    48
#> 6      1     6     2
#> 7      1     7     1
#> 8      1     8    18
#> 9      1     9     1
#> 10     1    10     2
#> 11     1    11   436
#> 12     1    12   528
#> 13     1    13    27
#> 14     1    14     1
#> 15     1    15     1
#> 16     1    16     1
#> 17     1    17     1
#> 18     1    18     1
#> 19     1    19     1
#> 20     1    20     1
#> 21     1    21     7

# ... from a binary raster, patches are automatically determined
mArea(obj = bin, scale = "patch")

#>    class patch cells
#> 1      1     1    98
#> 2      1     2    35
#> 3      1     3    25
#> 4      1     4   211
#> 5      1     5     1
#> 6      1     6     1
#> 7      1     7     1
#> 8      1     8     1
#> 9      1     9     3
#> 10     1    10     1
#> 11     1    11     8
#> 12     1    12   375
#> 13     1    13     1
#> 14     1    14     8
#> 15     1    15     1
#> 16     1    16     1
#> 17     1    17     8
#> 18     1    18     1
#> 19     1    19     1
#> 20     1    20     1
#> 21     1    21     6
#> 22     1    22   468
#> 23     1    23     1
#> 24     1    24     1
#> 25     1    25     1
#> 26     1    26     1

mArea(obj = bin, scale = "class")

#>   class cells
#> 1     0  2100
#> 2     1  1260