Edge length

Calculate the length of edges/boundaries between objects in a spatial raster.

mPerimeter(obj, scale = "patch", unit = "cells", layer = NULL)

Arguments

obj
[Raster*(1)]
The object to measure.

scale
[character(1)]
scale at which the area of objects should be calculated; possible values are "patch" and "class".

unit
[character(1)]
the unit the output should have. With "map" the result will be in the respective map unit and with "cells" (default) it will be the multiple of the raster cell dimension.

layer
[character(1)]
in case obj has several layers, specify here the layer for which the area of objects shall be calculated (by default, the first layer).

Value

For scale = "class" the edge length of each unique value (class) in the raster, for scale = "patch" the edge length of distinct objects per distinct values (i.e. the area of patches per class).

See also

Examples

cat <- rtData$categorical
bin <- rBinarise(rtData$continuous, thresh = 40)

# the perimeter ...
# ... per landcover type
mPerimeter(obj = cat, scale = "class")

#>   class edges
#> 1     1    34
#> 2    11   195
#> 3    21   153
#> 4    24   468
#> 5    27   249
#> 6    31    34
#> 7    41   646
#> 8    44   379
#> 9    47   426

# ... of patches per landcover typ
mPerimeter(obj = cat, scale = "patch")

#>    class patch edges
#> 1      1     1    34
#> 2     11     1   195
#> 3     21     1   153
#> 4     24     1   413
#> 5     24     2    55
#> 6     27     1   135
#> 7     27     2   114
#> 8     31     1    34
#> 9     41     1    34
#> 10    41     2    84
#> 11    41     3    76
#> 12    41     4    34
#> 13    41     5    66
#> 14    41     6     6
#> 15    41     7     4
#> 16    41     8    38
#> 17    41     9     4
#> 18    41    10     4
#> 19    41    11   134
#> 20    41    12    82
#> 21    41    13    80
#> 22    44     1    20
#> 23    44     2    28
#> 24    44     3     4
#> 25    44     4    64
#> 26    44     5     4
#> 27    44     6     6
#> 28    44     7    12
#> 29    44     8   124
#> 30    44     9    49
#> 31    44    10     6
#> 32    44    11    24
#> 33    44    12     6
#> 34    44    13    18
#> 35    44    14    14
#> 36    47     1   104
#> 37    47     2     6
#> 38    47     3   102
#> 39    47     4    26
#> 40    47     5     4
#> 41    47     6     4
#> 42    47     7     4
#> 43    47     8     4
#> 44    47     9     4
#> 45    47    10     4
#> 46    47    11     4
#> 47    47    12    12
#> 48    47    13   148

# ...  of certain values; from a binary raster, patches are 
# automatically determined
require(magrittr)
rBinarise(obj = cat, match = c(41, 44, 47)) %>%
  mPerimeter(scale = "patch", layer = "values_binarised")

#>    class patch edges
#> 1      1     1    34
#> 2      1     2    64
#> 3      1     3   122
#> 4      1     4    34
#> 5      1     5    66
#> 6      1     6     6
#> 7      1     7     4
#> 8      1     8    38
#> 9      1     9     4
#> 10     1    10     6
#> 11     1    11   182
#> 12     1    12   221
#> 13     1    13    26
#> 14     1    14     4
#> 15     1    15     4
#> 16     1    16     4
#> 17     1    17     4
#> 18     1    18     4
#> 19     1    19     4
#> 20     1    20     4
#> 21     1    21    12

mPerimeter(obj = bin, scale = "class")

#>   class edges
#> 1     0   714
#> 2     1   714