Obtain data from the European Atlas of Forest Tree Species

Obtain presence and habitat suitability data from the European Atlas of Forest Tree Species

oEFTA(mask = NULL, species = NULL, type = "rpp")

downloadEFTA(file = NULL, localPath = NULL)

Arguments

mask
[geom | SpatialPolygon*]
spatial object of which the extent is the area of interest.

species
[character(.)]
name(s) of species to get data for. Can be abbreviated if a catalog is provided.

type
[character(1)]
the dataset type, either "rpp" (relative probability of presence) or "mhs" (maximum habitat suitability).

file
[character(1)]
the name of the file to download.

localPath
[character(1)]
the local path where files are supposed to be stored (i.e. on your harddisc).

Value

bla

Details

https://w3id.org/mtv/FISE/map-data-RPP/v0-3-2/internet/* https://w3id.org/mtv/FISE/map-data-RPP/v0-3-2/internet/Acer-campestre

mention that values are integers and not "probabilities" between 0 and 1)

Functions

downloadEFTA: function to download data related to the EMMA dataset

References

de Rigo, D., Caudullo, G., Houston Durrant, T., San-Miguel-Ayanz, J., 2016. The European Atlas of Forest Tree Species: modelling, data and information on forest tree species. In: San-Miguel-Ayanz, J., de Rigo, D., Caudullo, G., Houston Durrant, T., Mauri, A. (Eds.), European Atlas of Forest Tree Species. Publ. Off. EU, Luxembourg, pp.

See also

Examples

# NOT RUN {
require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) %>%
  geomRectangle() %>%
  setCRS(crs = projs$laea)

# extract data for the derived mask
myTrees <- oEFTA(mask = myMask,
                 species = c("Quercus robur", "Alnus incana", "Betula sp"))
visualise(gridded = myTrees$`Betula sp`, trace = TRUE)
# }