Obtain data from the Atlas of European Mammals.

Obtain occurence data of mammals in Europe of the first EMMA dataset.

oEMMA(mask = NULL, species = NULL, version = 1, ...)

downloadEMMA(file = NULL, localPath = NULL)

Arguments

mask
[geom | Spatial* | sf]
spatial object of which the extent is the area of interest.

species
[character(.)]
name(s) of species to get occurrence for. Can be abbreviated if a catalog is provided.

version
[integerish(1)]
The version of the Atlas of European Mammals; the recent version is 1, but an update is in the making.

...
[various]
other arguments.

file
[character(1)]
the name of the file to download.

localPath
[character(1)]
the local path where files are supposed to be stored (i.e. on your harddisc).

Value

A data-frame of the species of interest occuring in the area outlined by mask.

Details

The object provided in mask is treated as a single mask, irrespective of that object consisting of only one or several features. The extent comprising all features (point(s), line(s), polygon(s)) is used as area of interest. This is in contrast to obtain, where a mask may consist of several features, each of which are treated as seperate mask.

The website Societas Europaea Mammalogica kindly offers maps of the ocurrence of all mammals in Europe, which are parsed by oEMMA. See load_svg for a technical explanation.

The occurrence data are recorded and stored on a grid of 50 x 50 km. This is the same grid as used for the Atlas Florae Europaeae.

The dataset meta_emma lists all available species.

Functions

downloadEMMA: function to download data related to the EMMA dataset

References

Mitchell-Jones A, Amori G, Bogdanowicz W, Kryštufek B, Reijnders P, Spitzenberger F, Stubbe M, Thissen J, Vohralík V and Zima J (1999). The Atlas of European Mammals. Academic Press, London

See also

Examples

# NOT RUN {
require(magrittr)

mySpecies <- oEMMA(mask = rtGeoms$mask,
                   species = c("Apodemus agrarius",
                               "Apodemus flavicollis",
                               "Vulpes vulpes"))

# Subset from a large set of files using an index
abbr_species <- function(x){
  tolower(paste(substr(unlist(strsplit(x, ' ')), 1, 3), collapse = '_'))
}
mySpecies <- catalog(path = rtPaths$emma$local,
                   type = 'svg', abbr = abbr_species) %>%
  subset(abbr %in% c("apo_agr", "apo_fla", "vul_vul")) %$%
  oEMMA(mask = myMask, species = original)

# get the (updated) bibliography
reference(style = "bibtex")
# }