Obtain MODIS data

Obtain data from various MODIS datasets.

oMODIS(mask = NULL, period = NULL, product = NULL, layer = NULL,
  process = TRUE)

downloadMODIS(file = NULL, localPath = NULL, getDates = NULL,
  getFiles = NULL)

Arguments

mask
[geom | Spatial*]
spatial object of which the extent is the area of interest.

period
[integerish(.)]
year of interest (YYYY), optionally concatenated to day of the year (DDD), which must be between 001 and 366.

product
[character(1)]
name of the modis-product of interest. Can be abbreviated if an index is provided.

layer
[character(1) | integerish(1)]
layer(s) of interest. Either the integer of the layers position or the name of the layer.

process
[logical(1)]
should the data be procesed further in R (TRUE, default) or should the data merely be downloaded and stored on the harddisc (FALSE)?

file
[character(1)]
the name of the file to download.

localPath
[character(1)]
the local path where files are supposed to be stored (i.e. on your harddisc).

getDates
[character(1)]
the online path to the directory where folders for all valid dates are stored (recently the first level after selecting the product).

getFiles
[character(1)]
the online path to the directory where product files are stored.

Value

A RasterLayer or RasterStack of MODIS data.

Details

The object provided in mask is treated as a single mask, irrespective of that object consisting of only one or several features. The extent comprising all features (point(s), line(s), polygon(s)) is used as area of interest. This is in contrast to obtain, where a mask may consist of several features, each of which are treated as seperate mask.

The period can either be given as a single value if only one year is of interest, or as two values, the start and end date. If you do not specify DDD, it will be set to "001" and "366" for start and end year respectively.

Supported MODIS-products - in the sense that predefined meta-data are available within this package - are:

MOD09A1: Surface Reflectance, 8 day, 500 meter, February 2000 - present

MOD09Q1: Surface Reflectance, 8 day, 250 meter, February 2000 - present

MOD11A2: Land Surface Temperature/Emissivity, 8 day, 1000 meter, March 2000 - present

MOD13Q1: Vegetation Indices, 16 day, 250 meter, February 2000 - present

MOD13A1: Vegetation Indices, 16 day, 500 meter, February 2000 - present

MOD13A2: Vegetation Indices, 16 day, 1000 meter, February 2000 - present

MOD13A3: Vegetation Indices, monthly, 1000 meter, February 2000 - present

MOD14A2: Thermal Anomalies and Fire, 8 day, 1000 meter, February 2000 - present

MOD15A2H: Leaf Area Index/FPAR, 8 day, 500 meter, July 2002 - present

MOD17A2H: Gross Primary Productivity, 8 day, 500 meter, February 2000 - present

MOD17A3H: Net Primary Production, yearly, 500 meter, 2000 - present

MOD44W: Land-Water Mask, yearly, 250 meter, 2000 - 2015

Get additional information about the spatial background from Marine Geospatial Ecology Tools

Functions

downloadMODIS: function to download data related to the MODIS products

See also

Examples

# NOT RUN {
water <- oMODIS(mask = rtGeoms$mask, period = 2000,
                product = "MOD44W", layer = 1)
visualise(gridded = water, trace = TRUE)

# not specifying 'layer' results in a stack of all layers
myMODIS <- oMODIS(mask = rtGeoms$mask, period = 2012,
                  product = "MOD17A3")
visualise(gridded = myMODIS, trace = TRUE)

modEvi <- oMODIS(mask = rtGeoms$mask, period = c(2016100, 2016250),
                 product = "MOD13A3", layer = "EVI")
visualise(gridded = modEvi, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")

# the modis tiles
modWindow <- data.frame(x = c(-20015109.354, 20015109.354),
                        y = c(-10007554.677, 10007554.677))
tiles_modis <- geomTiles(window = modWindow, cells = c(36, 18),
                         crs = projs$sinu)
visualise(geom = tiles_modis)
# }