Obtain global climate data

Obtain data from the 'worldclim' dataset (paper1, paper2).

oWCLIM(mask = NULL, variable = NULL, month = c(1:12), resolution = 0.5,
  version = 2)

downloadWCLIM(file = NULL, localPath = NULL)

Arguments

mask
[geom | SpatialPolygon*]
spatial object of which the extent is the area of interest.

variable
[character(.)]
the climatic variable of interest; see Details to find which variables are available in which version.

month
[integerish(.)]
the month(s) for which the data should be extracted.

resolution
[numeric(1)]
the spatial resolution in arc-minutes.

version
[integerish(1)]
the version for which you'd like to obtain the worldclim data.

file
[character(1)]
the name of the file to download.

localPath
[character(1)]
the local path where files are supposed to be stored (i.e. on your harddisc).

Details

Version 2.0 includes the variables:

"tmin": average monthly minimum temperature [°C * 10],

"tmax": average monthly maximum temperature [°C * 10],

"tavg": average monthly mean temperature [°C * 10],

"prec": average monthly precipitation [mm],

"srad": average monthly solar radiation [kJ m-2 day-1],

"wind": average monthly wind speed [m s-1],

"vapr": average monthly water vapour pressure [kPa],

"bio": bioclimatic variables derived from the above.

The files of version 1.4 are shipped in the 'bil' format. While the files are compressed into a zip-file, once unzipped, they take up a lot of space on the harddisc. Hence, when using this function for the first time, downloadWCLIM downloads and unzips the original file(s), transforms the monthly files into 'tif' format (which uses only about 1/10th of the space) and deletes the original 'bil' files. Beware that due to the size of the inital files this takes a lot (!) of time, especially for the 30 arc-second files. However, when repeatedly using these data, so the hope, this procedure should make the future work with these data smoother.

Functions

downloadWCLIM: function to download global climte datasets

See also

Examples

# NOT RUN {
require(magrittr)

# load and outline masks from file
myMask <- loadData(files = "aWindow.csv",
                   localPath = system.file("csv", package="rasterTools")) %>%
  geomRectangle() %>%
  setCRS(crs = projs$laea)

# extract worldclim data for the derived mask
myWCLIM <- oWCLIM(mask = myMask,
                  variable = c("tmin", "tmax"),
                  month = c(5:9))
visualise(gridded = myWCLIM$, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
# }