Obtain spatial datasets

Extract information from various spatial (gridded) datasets both from local and online resources.

obtain(data, mask)

Arguments

data
[list(.)]
algorithm in which the operators to load spatial datasets are specified. Each operator is a list iteself and includes the operator name and its arguments as sub-elements; see Examples.

mask
[geom | SpatialPolygon*]
spatial object of which the extent is the area of interest.

Value

Depends on the method called, but it is always organised in lists. The first hierarchical level of the list contains all the spatial units/masks. The second level contains an entry for each dataset and its temporal extent. The third level contains the resulting output. Mostly this would be an extracted raster according to what has been specified in the arguments, but for some operators the output is a data.frame or SpatialPointsDataFrame.

Details

obtain expects an individual directory for each dataset, where all the files of this particular dataset are located. rasterTools provides a list of paths getOption("rtPaths"), which contains the directories and urls of the local and online resources, respectively. In case the dataset files are not available locally and if a url is available, an attempt to download the files is undertaken. That means typically you do not have to use the dataset specific download* functions manually, they are documented nevertheless for those who are interested. In case an error occurs, it might be necessary to specify new paths in rtPaths (see updatePaths).

Recently supported datasets are:

oCLC: Corine Land Cover data

oEMMA: Occurence data of the Atlas of European Mammals

oGFC: Global Forest Change dataset

oMODIS: MODIS products

Examples

# NOT RUN {
require(magrittr)

# specify the datasets from which you want to get data
myDatasets <- list(list(operator = "oCLC", period = 2006),
                   list(operator = "oGFC", period = c(2005:2007)),
                   list(operator = "oMODIS", product = "mod17a3", period = 2006,
                        layer = 2))

# load and outline masks from a file with locations
myLocations <- loadData(files = "locations.csv",
                        localPath = system.file("csv", package="rasterTools"))
myMask <- gGroup(geom = myLocations, distance = 10000) %>%
  geomRectangle() %>%
  gToSp(crs = projs$laea)# %>%
  gBuffer(width = 1000, byid = TRUE)

# grab the data
myData <- obtain(data = myDatasets, mask = myMask)
# }