Generate spatial (gridded) patterns

This function is still largely work in process and very experimental (but will be more awesome than experiences of a similar kind!)

generate(model, dimensions)

Arguments

model
list of operators, in which the spatial patterns are specified. Each operator is a list iteself and includes the operator name and its arguments as sub-elements; see Examples.

dimensions
number of columns and rows the landscape model ought to have.

Details

Spatial pattern models, such as neutral landscape models (NLMs), are useful to generate distinct spatial patterns. Ideally, the resulting spatial patterns are strictly controlable. They help studying ecological processes or spatial statistics in general.

Many spatial pattern models (SPMs) have been suggested and neutral landscape models (NLMs) probably constitute the most common use-case. However, often they are merely treated as "habitat models", where only the presence of habitat is simulated. Various SPMs that depict "artificial patterns or processes" introduced by human disturbance or SPMs that depict the influence of geological, atmospheric and ecological dynamics are also of interest. Available are here ...

... neutral landscape models:

nlmGradient: Neutral landscape model based on a gradient.

nlmRandom: Randomly assembled neutral landscape model.

nlmHeightmap: Neutral landscape model based on a heightmap.

... other (spatial) pattern models:

pmNoise: Patterns based on different kinds of noise.

... ecological processes models:

epmSucces: Let objects in the spatial model success.

epmDiversify: Let objects in the spatial model diversify.

epmPerforate: Create gaps (e.g. clearcuts) into objects of the spatial model.

epmFragment: Create fragmentation in the spatial model.

epmConnect: Create connecting elements in the spatial model.

Combine these landscape models with spatial patterns (link to vignette here) and operators of modify to come up with a larger set of spatial models (link to vignette here).

Examples

# NOT RUN {
## neutral landscape models
nlm <- list(list(operator = "nlmRandom", seed = ),
            list(operator = "nlmHeightmap", fracdim = 2.3,
                 seed = 12769))

myLandscape <- generate(model = nlm, dimensions = c(300, 300), to_env = TRUE)
# }