Obtain data from the European Atlas of Forest Tree Species

Obtain presence and habitat suitability data from the European Atlas of Forest Tree Species

oEFTA(mask = NULL, species = NULL, type = "rpp")

downloadEFTA(file = NULL, localPath = NULL)

Arguments

mask
[geom | Spatial* | sf]
spatial object of which the extent is the area of interest.

species
[character(.)]
name(s) of species to get data for. Can be abbreviated if a catalog is provided.

type
[character(1)]
the dataset type, either "rpp" (relative probability of presence) or "mhs" (maximum habitat suitability).

file
[character(1)]
the name of the file to download.

localPath
[character(1)]
the local path where files are supposed to be stored (i.e. on your harddisc).

Value

bla

Details

The object provided in mask is treated as a single mask, irrespective of that object consisting of only one or several features. The extent comprising all features (point(s), line(s), polygon(s)) is used as area of interest. This is in contrast to obtain, where a mask may consist of several features, each of which are treated as seperate mask.

The values in this dataset are originally stored as values in the range of 0 and 1 (the probability of presence). Here they are transformed to integer factors from 0 to 100.

Functions

downloadEFTA: function to download data related to the EFTA dataset

References

de Rigo, D., Caudullo, G., Houston Durrant, T., San-Miguel-Ayanz, J., 2016. The European Atlas of Forest Tree Species: modelling, data and information on forest tree species. In: San-Miguel-Ayanz, J., de Rigo, D., Caudullo, G., Houston Durrant, T., Mauri, A. (Eds.), European Atlas of Forest Tree Species. Publ. Off. EU, Luxembourg, pp.

See also

Examples

# NOT RUN {
myTrees <- oEFTA(mask = rtGeoms$mask,
                 species = c("Quercus robur", "Alnus incana",
                             "Pinus sylvestris", "Betula sp"))
visualise(gridded = myTrees, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
# }