Obtain global climate data

Obtain data from the 'worldclim 2' dataset (paper).

oWCLIM(mask = NULL, variable = NULL, month = c(1:12),
  resolution = 0.5, ...)

downloadWCLIM(file = NULL, localPath = NULL)

Arguments

mask
[geom | Spatial* | sf]
spatial object of which the extent is the area of interest.

variable
[character(.)]
the climatic variable of interest; see Details.

month
[integerish(.)]
the month(s) for which the data should be extracted.

resolution
[numeric(1)]
the spatial resolution in arc-minutes.

...
[various]
other arguments.

file
[character(1)]
the name of the file to download.

localPath
[character(1)]
the local path where files are supposed to be stored (i.e. on your harddisc).

Details

The object provided in mask is treated as a single mask, irrespective of that object consisting of only one or several features. The extent comprising all features (point(s), line(s), polygon(s)) is used as area of interest. This is in contrast to obtain, where a mask may consist of several features, each of which are treated as seperate mask.

The following variables are included:

"tmin": average monthly minimum temperature [°C * 10],

"tmax": average monthly maximum temperature [°C * 10],

"tavg": average monthly mean temperature [°C * 10],

"prec": average monthly precipitation [mm],

"srad": average monthly solar radiation [kJ m-2 day-1],

"wind": average monthly wind speed [m s-1],

"vapr": average monthly water vapour pressure [kPa],

"bio": all bioclimatic variables,

"bio_01": annual mean temperature,

"bio_02": mean diurnal range,

"bio_03": Isothermality,

"bio_04": temperature seasonality,

"bio_05": Max temperature of warmest month,

"bio_06": Min temperature of coldest month,

"bio_07": temperature annual Range,

"bio_08": mean temperature of wettest quarter,

"bio_09": mean temperature of driest quarter,

"bio_10": mean temperature of warmest quarter,

"bio_11": mean temperature of coldest quarter,

"bio_12": annual precipitation,

"bio_13": precipitation of wettest month,

"bio_14": precipitation of driest month,

"bio_15": precipitation seasonality,

"bio_16": precipitation of wettest quarter,

"bio_17": precipitation of driest quarter,

"bio_18": precipitation of warmest quarter,

"bio_19": precipitation of coldest quarter.

Functions

downloadWCLIM: function to download global climte datasets

See also

Examples

# NOT RUN {
myWCLIM <- oWCLIM(mask = rtGeoms$mask, variable = c("tavg"),
                  month = c(5:9))
visualise(raster = myWCLIM, trace = TRUE)

# get the (updated) bibliography
reference(style = "bibtex")
# }